Update README.md
I have developed a project for generating Persian poetry using a compact GPT-like model with approximately 10.8 million parameters. While this model demonstrates potential, the training dataset was relatively small, which limited its ability to effectively capture long-term dependencies and nuanced representations in the text. As a result, the generated poetry did not meet my initial expectations.
To enhance the project's functionality, I have modularized the code and incorporated several cutting-edge techniques. This includes the implementation of various types of tokenizers—specifically, character-level, word-level, and subword-level tokenizers. These tokenizers allow the model to process and generate text more effectively. Additionally, I have employed advanced post-processing methods to improve the quality and coherence of the generated poetry.
Another significant challenge I faced was the limitation of computational resources for training the model. I utilized Google Colab for this purpose, which, while convenient, may not provide the optimal environment for extensive training.
I believe that with further training on more robust resources and a larger dataset, the quality of the generated text can be significantly improved. This project has the potential to evolve into a powerful tool for Persian poetry generation, and I am excited about the possibilities for future enhancements.